% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTs.R
\name{aggregateIndispo}
\alias{aggregateIndispo}
\title{Product ts from csv file}
\usage{
aggregateIndispo(path, outputFile, aggregatedCol, colDcast = NULL)
}
\arguments{
\item{path}{\code{character} csv file product by \code{\link{giveTS}}}

\item{outputFile}{\code{character} csv output file}

\item{aggregatedCol}{\code{character} column to aggregate, can be : "date", "hour", "month", "year", "productionType"
can also be a vector (c("month", "year"))}

\item{colDcast}{\code{character}, render result by columns : see exemple}
}
\description{
Product ts from csv file
}
\examples{

\dontrun{
 aggregateIndispo("NLindispoTS.csv", "agg1.csv", c("date", "hour", "productionType"))
 aggregateIndispo("NLindispoTS.csv", "agg2.csv",
  c("date", "hour", "productionType"), colDcast = "productionType")
}
}
