% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromEnsoe.R
\name{cleanAfterRead}
\alias{cleanAfterRead}
\title{Transform RDS file to csv}
\usage{
cleanAfterRead(dta)
}
\arguments{
\item{dta}{\code{data.table} data.table read with read.RDS}
}
\description{
Transform RDS file to csv
}
