% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataFromEnsoe.R
\name{getDataOutagesProduction}
\alias{getDataOutagesProduction}
\title{Load unavailability of production from Ensoe}
\usage{
getDataOutagesProduction(token, bz = "10YNL----------L", docType = "A80",
  docStatus = "A05", start = "2017-01-01", end = "2017-12-31",
  EIC = NULL, fileToMerge = "NLindispo.csv", by = "week",
  entsoeHttp = "https://transparency.entsoe.eu/api?")
}
\arguments{
\item{token}{\code{character}, ENTSOE token}

\item{bz}{\code{character} build zone, transparency.entsoe.eu API guide}

\item{docType}{\code{character} docType, A80 for generation unit and A77 for produciton unit}

\item{docStatus}{\code{character} docStatus "A05" : Active, "A09" : Cancel}

\item{start}{\code{character} start date}

\item{end}{\code{character} end date}

\item{EIC}{\code{character} optional, powerSystemResourcesmRID default NULL -> "All"}

\item{fileToMerge}{\code{character} csv where data will be save, you can also add information in an existing file.}

\item{by}{\code{character} load data by day, week, month, year?
Maximum of documents load by request is 200 so if data are not correctly try an other timestep.}

\item{entsoeHttp}{\code{character}, ENTSOE adress, defalut https://transparency.entsoe.eu/api?}
}
\description{
getDataOutagesProduction request ensoe API and create RDS file who contains unavailability of production.
}
\examples{
\dontrun{

 #PROD : "A77" Gene : "A80"
 token <- "35324aa6-f330-4240-ba21-81a1c7639548"
 #BiddingZone_Domain = "10YFR-RTE------C"
 #BiddingZone_Domain = "10YBE----------2"
 #BiddingZone_Domain = "10YNL----------L"
 #BiddingZone_Domain = "10Y1001A1001A63L"

 #BiddingZone_Domain = "10YDE-VE-------2"    #GRT Allemand 50Hertz
 #BiddingZone_Domain = "10YDE-RWENET---I"    #GRT Allemand Amprion
 #BiddingZone_Domain = "10YDE-EON------1"    #GRT Allemand TennetDE
 #BiddingZone_Domain = "10YDE-ENBW-----N"    #GRT Allemand TransnetBW
 #BiddingZone_Domain = "10YAT-APG------L"    #Autriche

 #NL
 getDataOutagesProduction(token = token, bz = "10YNL----------L",
  docType = "A80",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "NLindispo.csv")
 getDataOutagesProduction(token = token, "10YNL----------L", docType = "A77",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "NLindispo.csv", by = "year")

 #BE
 getDataOutagesProduction(token = token, bz = "10YBE----------2", docType = "A80",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "BEindispo.csv")
 getDataOutagesProduction(token = token, "10YBE----------2", docType = "A77",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "BEindispo.csv")

 #FR
 getDataOutagesProduction(token = token, bz = "10YFR-RTE------C", docType = "A80",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "FRindispo.csv")
 getDataOutagesProduction(token = token, "10YFR-RTE------C", docType = "A77",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "FRindispo.csv")

 #DE
 getDataOutagesProduction(bz = "10YDE-VE-------2", docType = "A80",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "DEindispo.csv")
 getDataOutagesProduction("10YDE-VE-------2", docType = "A77",
  start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv", by = "year")


 getDataOutagesProduction(token = token, bz = "10YDE-RWENET---I", docType = "A80",
   start = "2015-01-01", end = "2018-12-31", fileToMerge = "DEindispo.csv")
 getDataOutagesProduction(token = token, bz = "10YDE-RWENET---I", docType = "A77",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv", by = "year")

 getDataOutagesProduction(token = token, bz = "10YDE-EON------1", docType = "A80",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv", by = "month")

 ##Week (for +200 row)
 getDataOutagesProduction(token = token, bz = "10YDE-EON------1", docType = "A80",
   start = "2015-08-01", end = "2015-09-10", fileToMerge = "DEindispo.csv")
 getDataOutagesProduction(token = token, "10YDE-EON------1", docType = "A80",
   start = "2016-06-01", end = "2016-09-10", fileToMerge = "DEindispo.csv")
 getDataOutagesProduction(token = token, bz = "10YDE-EON------1", docType = "A80",
   start = "2016-07-27", end = "2016-08-03", fileToMerge = "DEindispo.csv", by = "day")
 getDataOutagesProduction(token = token, bz = "10YDE-EON------1", docType = "A80",
   start = "2016-12-01", end = "2017-01-10", fileToMerge = "DEindispo.csv")
 getDataOutagesProduction(token = token, bz = "10YDE-EON------1", docType = "A80",
   start = "2017-06-01", end = "2017-09-10", fileToMerge = "DEindispo.csv")

 getDataOutagesProduction(token = token, bz = "10YDE-EON------1", docType = "A77",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv")

 getDataOutagesProduction(token = token, bz = "10YDE-ENBW-----N", docType = "A80",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv", by = "month")
 getDataOutagesProduction(token = token, bz = "10YDE-ENBW-----N", docType = "A77",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv", by = "year")

 getDataOutagesProduction(token = token, bz = "10YAT-APG------L", docType = "A80",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv", by = "month")
 getDataOutagesProduction(token = token, bz = "10YAT-APG------L", docType = "A77",
   start = "2015-01-01", end = "2019-01-01", fileToMerge = "DEindispo.csv", by = "year")


}
}
