% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNetTransferCapacity.R
\name{getEstimedNetTransferCapacity}
\alias{getEstimedNetTransferCapacity}
\title{Load Estimed Net Transfer Capacity from ENTSOE}
\usage{

  getEstimedNetTransferCapacity(entsoeHttp = "https://transparency.entsoe.eu/api?",
  ...)
}
\arguments{
\item{entsoeHttp}{\code{character} enstoe API defalut : "https://transparency.entsoe.eu/api?"}

\item{...}{others params}
}
\description{
getDataOutagesProduction request ensoe API.
}
\examples{
\dontrun{

getEstimedNetTransferCapacity(
  securityToken = "3ea2c90d-8ecb-4452-898a-263ea835f498",
  documentType="A61",
  contract_MarketAgreement.Type = "A01",
  in_Domain = "10YFR-RTE------C",
  out_Domain = "10YGB----------A",
  periodStart = "2018-10-12",
  periodEnd = "2018-10-12")
}

}
