% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTs.R
\name{giveTS}
\alias{giveTS}
\title{Transform RDS file to csv}
\usage{
giveTS(path, outFile, minDate = as.POSIXct("2015-01-01", tz = "UTC"),
  maxDate = as.POSIXct("2018-12-31", tz = "UTC"), overlappingvalues = "max",
  clipping = FALSE)
}
\arguments{
\item{path}{\code{character} data path (refear to a CSV file)}

\item{outFile}{\code{character} file to write (.csv)}

\item{minDate}{\code{POSIXct} minimum of dates}

\item{maxDate}{\code{POSIXct} maximum of dates}

\item{overlappingvalues}{\code{character} operation for overlap, can be "sum" or "min"}

\item{clipping}{\code{boolean}, clipping times series by maximum capacity of produciton}
}
\description{
Transform RDS file to csv
}
\examples{
\dontrun{
  giveTS("FRindispo2.csv", outFile = "FRindispoTS.csv")
}

}
