% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{runIndispoApp}
\alias{runIndispoApp}
\title{Load unavailability of production from Ensoe}
\usage{
runIndispoApp(path)
}
\arguments{
\item{path}{\code{character} csv file}
}
\description{
getDataOutagesProduction request ensoe API and create RDS file who contains unavailability of production.
}
\examples{
\dontrun{
  runIndispoApp("FFFindispoTEST8006.csv")
}

}
