% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{unavailabilityPUTS}
\alias{unavailabilityPUTS}
\title{unavailability time serie after aggregation}
\usage{
unavailabilityPUTS(path, name)
}
\arguments{
\item{path}{\code{character} csv file}

\item{name}{\code{character} powerSystemResourcesName(s)}
}
\description{
Time serie vizualisation of unavailability
}
\examples{
\dontrun{


getDataOutagesProduction(token = token, bz = "10YBE----------2", docType = "A80",
                         start = "2018-10-01", end = "2018-12-31", fileToMerge = "BE.csv")


getDataOutagesProduction(token = token, "10YBE----------2", docType = "A77",
                         start = "2018-10-01", end = "2018-12-31", fileToMerge = "BE.csv")



giveTS("BE.csv", outFile = "BETS.csv", overlappingvalues = "sum")
aggregateIndispo("BETS.csv", "BEAGG.csv",
  c("date", "hour", "productionType"), colDcast = "productionType")
  
  
unavailabilityPUTS("BEAGG.csv", c("Fossil Gas", "Nuclear"))


}

}
