% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{unavailabilityTS}
\alias{unavailabilityTS}
\title{unavailability time serie for generation unit or produciton unit}
\usage{
unavailabilityTS(path, name)
}
\arguments{
\item{path}{\code{character} csv file}

\item{name}{\code{character} powerSystemResourcesName(s)}
}
\description{
Time serie vizualisation of unavailability
}
\examples{
\dontrun{


getDataOutagesProduction(token = token, bz = "10YBE----------2", docType = "A80",
                         start = "2018-10-01", end = "2018-12-31", fileToMerge = "BE.csv")


getDataOutagesProduction(token = token, "10YBE----------2", docType = "A77",
                         start = "2018-10-01", end = "2018-12-31", fileToMerge = "BE.csv")



giveTS("BE.csv", outFile = "BETS.csv", overlappingvalues = "sum")
unavailabilityTS("BETS.csv", c("PU.DOEL 1", "PU.DOEL 2", "PU.DOEL 4"))

}

}
